+++ test/src/factories/position_factory.dart
@@ -5,6 +5,8 @@ class PositionFactory {
     return Position(
       latitude: 52.561270,
       longitude: 5.639382,
+      timestamp: DateTime.fromMillisecondsSinceEpoch(500, isUtc: true),
+      mocked: true,
       altitude: 3000.0,
       accuracy: 0.0,
       heading: 0.0,